requirements.txt
aiohttp
click==8.3.1
fal-client
ultralytics==8.3.153
torch==2.8.0
torchvision==0.23.0
numpy==1.26.4
opencv-python
pillow==12.0.0
transformers==4.45.0
flask
